Move checkExistingAddresses to helpers/utils/util and stop its use of ramda (#8868)
