+++ ui/app/pages/add-token/token-list/token-list.component.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import classnames from 'classnames'
-import { checkExistingAddresses } from '../util'
+import { checkExistingAddresses } from '../../../helpers/utils/util'
 import TokenListPlaceholder from './token-list-placeholder'
 
 export default class InfoBox extends Component {
