+++ ui/app/pages/send/send-content/add-recipient/add-recipient.js
@@ -5,9 +5,7 @@ import {
   INVALID_RECIPIENT_ADDRESS_NOT_ETH_NETWORK_ERROR,
 } from '../../send.constants'
 
-import { isValidAddress, isEthNetwork } from '../../../../helpers/utils/util'
-import { checkExistingAddresses } from '../../../add-token/util'
-
+import { isValidAddress, isEthNetwork, checkExistingAddresses } from '../../../../helpers/utils/util'
 import ethUtil from 'ethereumjs-util'
 import contractMap from 'eth-contract-metadata'
